REFACTOR: 범용성을 위해 toast hook을 함수로 리팩터링
일반적인 함수로 제공되는 것을 hook으로 구현했었는데 이럴 경우 사용이 귀찮고 사용에 제약이 있을 것 같다고 판단하여 함수로 변경
